Account Page Updates
Fixed issues on account page and added more functionality like: being able to view invoices, able to pay for due invoices, able to change email and password, and able to change the auto renew status of the account. All forms on account path are now validated properly on the client side.