User Update Bug Fix
Update Current User now takes new token that the server constructs, decodes it, and sets it as the new user when the user updates something like their payment methods. Data is now persistant even if the browser refreshes, and the application is more secure as a whole due to tokenization. Various other visual fixes were made as well.