Nodemailer & Password Reset
Nodemailer was integrated more to successfully reset the user's password.  Front-end component to reset the password was also fixed. Nodemailer was also integrated to send the user an email when they make a purchase. For this, mustache (hogan.js) was used for email templating. Other nodemailer routes like quoting and invoice generated notifications still need to be integrated with nodemailer.