Initial Release Build
Invoicing system, admin authentication, header overflow, user request authorization through the request header, administrator functions, proper notifications for login and more notifications where all fixed. This is the initial release build that is able to be securely deployed to the Web Dream Domain. There is much more work to be done, but at this point, Web Dream is READY TO RELEASE!!!!